fix - 투표가 존재하지 않을시 에러가 발생

+++ src/main/kotlin/com/yareach/voting_system/election/service/ElectionService.kt
@@ -38,6 +38,9 @@ class ElectionServiceImpl(
     }
 
     override suspend fun deleteElection(id: String) {
+        if (!electionRepository.isExists(id)) {
+            throw ApiException(ErrorCode.ELECTION_NOT_FOUND, "electionId: $id")
+        }
         electionRepository.deleteById(id)
     }
